requirements.txt
numpy
pyproj
carla
pytz
pysolar
pandas
pvlib
matplotlib
google-genai
requests
lxml
